Add tool for measuring CPU core to core RTT through CPU caches
New tool to measure the RTT between CPU cores in a system.
Ping-pongs a cache line across a pair of CPU cores and uses an
NTP type algorithm to measure the actual round trip time. Designed
for x86 for now through the use of rdtsc instruction and
assumptions around x86 cache implementations